outputs.postgresql: Use a deterministic tag key
The previous code used a postgres serial for the tag table key. This results in telegraf having to lookup from the table to obtain the key. While there was an LRU cache to keep telegraf from having to constantly query, this posed a few problems.
1. The LRU cache is of a fixed size, but the performance can go from perfectly fine, to rock bottom as soon as you exceed the cache size. There is no gradual degredation that would allow the admin to detect the issue before it gets too bad. This has major repercussions as telegraf won't be able to keep up with ingest rate and will start dropping data.
2. The default for this cache was really small (1000), which is even likely too small for common use cases.
3. When telegraf starts up, the cache is cold, and so performance will be terrible while it is repopulated, again resulting in data loss.

The new solution solves this by using a deterministic tag key that does not require a lookup. The key is a 64-bit integer from the hash/maphash package. This hash is specifically designed for creating a hash of a string/bytes that can be used as a lookup key.