outputs.postgresql: Sort columns on table creation
This is just a convenience so that when doing `select *`, you don't get a random order. This is not meant to be a perfect solution as any additional columns that are added to the table definition will not be sorted (this is just how postgres works). It's just meant so that the column ordering isn't a complete mess, making the life of the DBA somewhat better.