outputs.postgresql: use connection pool
This switches the output to use a connection pool for managing the connection. Previously connection management was being done manually, accomplishing the same thing as if you had a pool size of 1. So this makes it much simpler.

This also removes the dbWrapper abstraction layer. There's no point to it when there's only one database driver. Plus having access to the underlying pool will make future work easier (without having to expand the abstraction, at which point it's no longer abstracting).