outputs.postgresql: don't try to detect mismatched data types
the admin may deliberately re-type a column because because they know a value will never exceed a more optimized type. The error response is the same in both cases (permanent with sub-batch drop), so let the DB handle it instead.
In the future, we may want to detect values that won't fit inside a column type on a per-row basis (in TableSource), but this means more code, and right now handling the admin's broken database scheme isn't worth it.