outputs.postgresql: Add handling for table & column name overflow
postgresql identifiers have a max length of 63 bytes. Previously if we tried to create a table or column with a name exceeding that, postgres would truncate it and proceed with the truncated value. This causes problems for us as we don't know the truncated value. And on top of that, when using tag tables, the truncated tag table name will be the same as the normal table name and result in a conflict.

To address the problem, we simply don't allow using table or column names over 63 characters. We same rules as schema modification statements being disabled. Meaning for a field we drop the bad field and insert the rest. For a tag we drop the whole metric.